Se agrega archivo con configuración de DR Sumar
